# cmd/swag/docs/swagger.yaml
basePath: /api/v1
definitions:
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  response.UserDto:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      password:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8082
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Test Swagger Example API
  version: "1.0"
paths:
  /api/v1/user/getUser:
    post:
      consumes:
      - application/json
      description: register
      parameters:
      - description: Email of the user
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Show an account
      tags:
      - accounts
  /api/v1/user/login:
    post:
      consumes:
      - application/json
      description: register
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Show an account
      tags:
      - accounts
  /api/v1/user/register:
    post:
      consumes:
      - application/json
      description: register
      parameters:
      - description: payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/response.UserDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Show an account
      tags:
      - accounts
swagger: "2.0"