# modulB/src/dto/response.go
package dto

import "encoding/json"

type Response struct {
	Success    bool `json:"success"`
	Message    string `json:"message"`
	Code       int16 `json:"code"`
	Timestamp  int64`json:"timestamp"`
	Result     json.RawMessage `json:"result"`
}